requirements.txt
Flask==2.3.2
gunicorn==20.1.0
flask_sqlalchemy==3.0.5
flask_migrate==4.0.4
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.5.5
python-dotenv==1.0.0
flask-cors==5.0.0
moviepy==1.0.3
Pillow==9.5.0
requests==2.31.0
# openai-whisper==20231117
assemblyai==0.33.0
flower==2.0.0